% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configuration Function}
\usage{
config(data_file)
}
\arguments{
\item{data_file}{the name (in quotes "") of the data frame undergoing QC}
}
\value{
List of paths and values
}
\description{
Configuration Function
}
