% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{logs}
\alias{logs}
\title{Initialize logging configuration}
\usage{
logs(log_file = "logs.txt")
}
\arguments{
\item{log_file}{The name of the log file to use}
}
\value{
list of logging functions
}
\description{
Initializes logging configuration with log file setup and logging functions
}
