% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_main.R
\name{qc_main}
\alias{qc_main}
\title{Quality Control Process}
\usage{
qc_main(registry_data, rules, yml_path = getwd(), log_file = "log/logs.txt")
}
\arguments{
\item{yml_path}{Directory path where YAML and HTML reports will be saved}

\item{log_file}{Path to the log file where logs will be stored}

\item{data_file}{Name of the data file to be validated}

\item{rules_file_path}{Path to the rules Excel file}
}
\value{
agent Object containing the results and configuration of the validation process
}
\description{
Quality Control Process
}
