% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_failed.R
\name{remove_failed}
\alias{remove_failed}
\title{Remove failed data}
\usage{
remove_failed(agent, data, rule)
}
\arguments{
\item{agent}{Run qc_main to retrieve agent}

\item{data}{Registry data}

\item{rule}{Rule #}
}
\description{
Remove failed data
}
