% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_main.R
\name{qc_main}
\alias{qc_main}
\title{Main QC check file}
\usage{
qc_main(data_file, rules_file_path)
}
\arguments{
\item{data_file}{takes in data set}

\item{rules_file_path}{takes in the rules file}
}
\value{
agent
}
\description{
Main QC check file
}
